pyproject.toml
[project]
name = "deputy"
version = "0.1.0"
description = "Agentic bot to help development engineers handle bugs and prioritize tasks"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.9.0",
    "websockets>=12.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "sentry-sdk>=1.38.0",
    "pygithub>=1.59.0",
    "aiofiles>=23.2.0",
    "langgraph>=0.0.69",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.8",
    "langchain-anthropic>=0.1.0",
    "pillow>=10.0.0",
    "requests>=2.31.0",
]
